examples/eoapi-auth-utils/pyproject.toml
[project]
dependencies = [
  "uvicorn>=0.27.1",
  "eoapi-auth-utils>=0.1.0",
]
description = "Example client built with eoapi-auth-utils"
name = "eoapi-auth-utils-example-client"
requires-python = ">=3.9"
version = "0.4.1"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.hatch.build.targets.wheel]
packages = ["app"]
